requirements.txt
ultralytics
streamlit
pillow